--- packages/vscode-webui/src/components/prompt-form/auto-completion/extension.tsx
@@ -1,11 +1,0 @@
-import Mention from "@tiptap/extension-mention";
-import { PluginKey } from "@tiptap/pm/state";
-
-export const autoCompletePluginKey = new PluginKey("autoCompletePluginKey");
-
-/**
- * A custom TipTap extension to handle mentions (like @name).
- */
-export const AutoCompleteExtension = Mention.extend({
-  name: "autoCompletion",
-});
