+++ packages/vscode-webui/src/lib/fuzzy-search.ts
@@ -3,7 +3,7 @@ import uFuzzy from "@leeoniya/ufuzzy";
 const MaxResult = 500;
 
 // Create a single uFuzzy instance to reuse
-export const ufInstance = new uFuzzy({
+const ufInstance = new uFuzzy({
   // Allow @ symbols and basic path characters
   intraChars: "[a-z\\d'@\\-_./]",
   // Don't split on forward slashes - treat them as regular characters
